fix SignUp and admni page after last merge
