--- resources/js/pages/Admin/pages/Serials.js
@@ -1,23 +1,0 @@
-//страница редактирования сериалов
-// Вывод списка сериалов, поиск по названию и году, удаление.
-
-
-import { ParserSerials } from '../../../components/Parser';
-
-export const SerialsAdmin = () => {
-    return (
-        <div>
-            <h2>Редактирование сериалов</h2>
-            <br />
-            <ParserSerials />
-        </div>
-    );
-};
-
-
-
-
-
-
-
-
